The InputDrawAdapter flattens the Json input into a list of InputDraw objects.
The MojoView(CustomView) can directly draws the objects in the flattened(adapted) Input list.

For example:
Json:
{
	"background_color": "#000000",
	"padding": 0.25,
	"children": [
		{
			"background_color": "#ff0000",
			"anchor_x": "left",
                	"anchor_y": "bottom"
		}
	]

}

is adapted(flattened) into a kotlin List<InputDraw>:
listOf(
	InputDraw(Bounds(0,0, 1000, 1000), "#000000"),
	InputDraw(Bounds(250, 250, 750 750), "#000000"),
)

The MojoView can iterate the flattened list and draw each InputDraw object. The MojoView is now lightweight is it just draws rectangle with a bound and a color.

The InputDrawAdapter is also unit testable, we can easily detect regressions in the bounds calculation, and also add new tests for new fields for example if we add a "margin" to the json.