Added support for media and media_content_mode.
Picasso is used for image loading.
For media_content_mode fill Picasso handles the image scaling.
For media_content_mode fit Picasso loads the image but the scaling is calculated.

Since Picasso runs on a background thread, I've added coroutines. So Picasso can load images in the background, but now all the RenderObject generation is done on the background.
I had to add Espresso IdlingResource for coroutines so that the tests wait until Picasso loads an image, and then after assert if the images are the same.